Added export tests
