+++ .gitignore
@@ -1,1 +1,2 @@
 rust/target/
+rust/hanb.hsit
